pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "drone_detector_mlops"
version = "0.0.1"
description = "A drone detection framework with extra MLOps sauce"
authors = [
  { name = "dtu-mlops", email = "your@email.com" },
]
license = { file = "LICENSE" }
keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dvc[gs]>=3.65.0",
    "fastapi==0.115.6",
    "kaggle>=1.8.3",
    "matplotlib>=3.10.8",
    "numpy>=2.4.0",
    "pillow>=12.1.0",
    "scikit-learn>=1.8.0",
    "timm>=1.0.23",
    "torch==2.6.0",
    "torchvision>=0.21.0",
    "typer==0.15.1",
    "uvicorn==0.34.0",
]

[dependency-groups]
dev = [
    "coverage==7.6.9",
    "invoke==2.2.0",
    "mkdocs-material==9.4.6",
    "mkdocs==1.6.1",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff>=0.5.3",
    "mkdocstrings-python==1.12.2",
    "ipython>=9.8.0",
    "jupyter>=1.1.1",
    "prek>=0.2.25",
]

[tool.ruff]
line-length = 120

[tool.coverage.run]
omit = ["tests/*"]
